moved last helper to helper.js & tested
